# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Project {
  id: number;
  name: string;
  language: string;
  createdAt: string;
  updatedAt: string;
  lastOpenedAt?: string | null;
  fileCount?: number;
}

export interface CreateProjectBody {
  name: string;
  language?: string;
}

export interface UpdateProjectBody {
  name?: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  name: string;
  path: string;
  content?: string | null;
  isFolder: boolean;
  parentId?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateFileBody {
  name: string;
  content?: string;
  isFolder?: boolean;
  parentId?: number | null;
}

export interface UpdateFileBody {
  name?: string;
  content?: string;
}

export interface RunCodeBody {
  code: string;
  language?: string;
}

export interface ExecutionResult {
  stdout: string;
  stderr: string;
  exitCode: number;
  executionTime: number;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface AiChatBody {
  message: string;
  code?: string | null;
  conversationHistory?: ChatMessage[];
}

export interface AiChatResponse {
  message: string;
  role: string;
}

export interface ActivityItem {
  type: string;
  description: string;
  timestamp: string;
}

export interface DashboardStats {
  totalProjects: number;
  totalFiles: number;
  totalExecutions: number;
  recentActivity: ActivityItem[];
}