# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: files
    description: File management
  - name: execution
    description: Code execution
  - name: ai
    description: AI assistant
  - name: dashboard
    description: Dashboard data
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectBody"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectBody"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/files:
    get:
      operationId: listFiles
      tags: [files]
      summary: List files in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"
    post:
      operationId: createFile
      tags: [files]
      summary: Create a file or folder
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFileBody"
      responses:
        "201":
          description: Created file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"

  /projects/{projectId}/files/{fileId}:
    get:
      operationId: getFile
      tags: [files]
      summary: Get file content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: File with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
    patch:
      operationId: updateFile
      tags: [files]
      summary: Update file name or content
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFileBody"
      responses:
        "200":
          description: Updated file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectFile"
    delete:
      operationId: deleteFile
      tags: [files]
      summary: Delete a file
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: fileId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{projectId}/run:
    post:
      operationId: runCode
      tags: [execution]
      summary: Execute code in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RunCodeBody"
      responses:
        "200":
          description: Execution result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExecutionResult"

  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Send a message to the AI assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/recent:
    get:
      operationId: getRecentProjects
      tags: [dashboard]
      summary: Get recently opened projects
      responses:
        "200":
          description: Recent projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        language:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        lastOpenedAt:
          type: string
          format: date-time
          nullable: true
        fileCount:
          type: integer
      required:
        - id
        - name
        - language
        - createdAt
        - updatedAt

    CreateProjectBody:
      type: object
      properties:
        name:
          type: string
        language:
          type: string
          default: python
      required:
        - name

    UpdateProjectBody:
      type: object
      properties:
        name:
          type: string

    ProjectFile:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        name:
          type: string
        path:
          type: string
        content:
          type: string
          nullable: true
        isFolder:
          type: boolean
        parentId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - name
        - path
        - isFolder
        - createdAt
        - updatedAt

    CreateFileBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string
          default: ""
        isFolder:
          type: boolean
          default: false
        parentId:
          type: integer
          nullable: true
      required:
        - name

    UpdateFileBody:
      type: object
      properties:
        name:
          type: string
        content:
          type: string

    RunCodeBody:
      type: object
      properties:
        code:
          type: string
        language:
          type: string
          default: python
      required:
        - code

    ExecutionResult:
      type: object
      properties:
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        executionTime:
          type: number
      required:
        - stdout
        - stderr
        - exitCode
        - executionTime

    AiChatBody:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
          nullable: true
        conversationHistory:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - message

    ChatMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
      required:
        - role
        - content

    AiChatResponse:
      type: object
      properties:
        message:
          type: string
        role:
          type: string
      required:
        - message
        - role

    DashboardStats:
      type: object
      properties:
        totalProjects:
          type: integer
        totalFiles:
          type: integer
        totalExecutions:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/ActivityItem"
      required:
        - totalProjects
        - totalFiles
        - totalExecutions
        - recentActivity

    ActivityItem:
      type: object
      properties:
        type:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - type
        - description
        - timestamp