# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all projects
 */
export const ListProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastOpenedAt: zod.coerce.date().nullish(),
  fileCount: zod.number().optional(),
});
export const ListProjectsResponse = zod.array(ListProjectsResponseItem);

/**
 * @summary Create a new project
 */
export const createProjectBodyLanguageDefault = `python`;

export const CreateProjectBody = zod.object({
  name: zod.string(),
  language: zod.string().default(createProjectBodyLanguageDefault),
});

/**
 * @summary Get a project by ID
 */
export const GetProjectParams = zod.object({
  projectId: zod.coerce.number(),
});

export const GetProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastOpenedAt: zod.coerce.date().nullish(),
  fileCount: zod.number().optional(),
});

/**
 * @summary Update a project
 */
export const UpdateProjectParams = zod.object({
  projectId: zod.coerce.number(),
});

export const UpdateProjectBody = zod.object({
  name: zod.string().optional(),
});

export const UpdateProjectResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastOpenedAt: zod.coerce.date().nullish(),
  fileCount: zod.number().optional(),
});

/**
 * @summary Delete a project
 */
export const DeleteProjectParams = zod.object({
  projectId: zod.coerce.number(),
});

/**
 * @summary List files in a project
 */
export const ListFilesParams = zod.object({
  projectId: zod.coerce.number(),
});

export const ListFilesResponseItem = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  path: zod.string(),
  content: zod.string().nullish(),
  isFolder: zod.boolean(),
  parentId: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListFilesResponse = zod.array(ListFilesResponseItem);

/**
 * @summary Create a file or folder
 */
export const CreateFileParams = zod.object({
  projectId: zod.coerce.number(),
});

export const createFileBodyContentDefault = ``;
export const createFileBodyIsFolderDefault = false;

export const CreateFileBody = zod.object({
  name: zod.string(),
  content: zod.string().default(createFileBodyContentDefault),
  isFolder: zod.boolean().default(createFileBodyIsFolderDefault),
  parentId: zod.number().nullish(),
});

/**
 * @summary Get file content
 */
export const GetFileParams = zod.object({
  projectId: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

export const GetFileResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  path: zod.string(),
  content: zod.string().nullish(),
  isFolder: zod.boolean(),
  parentId: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update file name or content
 */
export const UpdateFileParams = zod.object({
  projectId: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

export const UpdateFileBody = zod.object({
  name: zod.string().optional(),
  content: zod.string().optional(),
});

export const UpdateFileResponse = zod.object({
  id: zod.number(),
  projectId: zod.number(),
  name: zod.string(),
  path: zod.string(),
  content: zod.string().nullish(),
  isFolder: zod.boolean(),
  parentId: zod.number().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a file
 */
export const DeleteFileParams = zod.object({
  projectId: zod.coerce.number(),
  fileId: zod.coerce.number(),
});

/**
 * @summary Execute code in a project
 */
export const RunCodeParams = zod.object({
  projectId: zod.coerce.number(),
});

export const runCodeBodyLanguageDefault = `python`;

export const RunCodeBody = zod.object({
  code: zod.string(),
  language: zod.string().default(runCodeBodyLanguageDefault),
});

export const RunCodeResponse = zod.object({
  stdout: zod.string(),
  stderr: zod.string(),
  exitCode: zod.number(),
  executionTime: zod.number(),
});

/**
 * @summary Send a message to the AI assistant
 */
export const AiChatBody = zod.object({
  message: zod.string(),
  code: zod.string().nullish(),
  conversationHistory: zod
    .array(
      zod.object({
        role: zod.enum(["user", "assistant"]),
        content: zod.string(),
      }),
    )
    .optional(),
});

export const AiChatResponse = zod.object({
  message: zod.string(),
  role: zod.string(),
});

/**
 * @summary Get dashboard summary stats
 */
export const GetDashboardStatsResponse = zod.object({
  totalProjects: zod.number(),
  totalFiles: zod.number(),
  totalExecutions: zod.number(),
  recentActivity: zod.array(
    zod.object({
      type: zod.string(),
      description: zod.string(),
      timestamp: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Get recently opened projects
 */
export const GetRecentProjectsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  language: zod.string(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
  lastOpenedAt: zod.coerce.date().nullish(),
  fileCount: zod.number().optional(),
});
export const GetRecentProjectsResponse = zod.array(
  GetRecentProjectsResponseItem,
);