# lib/api-zod/src/generated/types/aiChatBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface AiChatBody {
  message: string;
  code?: string | null;
  conversationHistory?: ChatMessage[];
}