# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from "./activityItem";

export interface DashboardStats {
  totalProjects: number;
  totalFiles: number;
  totalExecutions: number;
  recentActivity: ActivityItem[];
}