# lib/api-zod/src/generated/types/projectFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectFile {
  id: number;
  projectId: number;
  name: string;
  path: string;
  content?: string | null;
  isFolder: boolean;
  parentId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}